requirements.txt.txt
python-telegram-bot==20.3
gspread==6.0.0
google-auth==2.23.4
nest_asyncio==1.5.8